fix(log): complete transition logging
