+++ pkg/plugins/runtime/k8s/webhooks/dataplane_validator.go
@@ -64,11 +64,9 @@ func (h *DataplaneValidator) ValidateCreate(ctx context.Context, req admission.R
 	if err := h.validator.ValidateCreate(ctx, core_model.MetaToResourceKey(coreRes.GetMeta()), coreRes, mesh); err != nil {
 		if kumaErr, ok := err.(*validators.ValidationError); ok {
 			resp := convertSpecValidationError(kumaErr, false, k8sRes)
-			logWebhookRejection(req, resp)
 			return resp
 		}
 		resp := admission.Denied(err.Error())
-		logWebhookRejection(req, resp)
 		return resp
 	}
 	return admission.Allowed("")
@@ -101,11 +99,9 @@ func (h *DataplaneValidator) ValidateUpdate(ctx context.Context, req admission.R
 	if err := h.validator.ValidateUpdate(ctx, coreRes, mesh); err != nil {
 		if kumaErr, ok := err.(*validators.ValidationError); ok {
 			resp := convertSpecValidationError(kumaErr, false, k8sRes)
-			logWebhookRejection(req, resp)
 			return resp
 		}
 		resp := admission.Denied(err.Error())
-		logWebhookRejection(req, resp)
 		return resp
 	}
 	return admission.Allowed("")
